added shared module for ui stuff, moved spinner there

+++ ui/src/app/admin/collection/photo-details/photo-details.component.ts
@@ -60,6 +60,9 @@ export class PhotoDetailsComponent implements OnInit, OnDestroy {
   }
 
   delete(photo: Photo): void {
+    if (!confirm('Delete photo?')) {
+      return;
+    }
     this.photoService.delete(this.collection, photo);
     this.router.navigate(['admin', 'collection', this.collection.slug]);
   }
